move jinja to substitutions module
