+++ esphome/components/substitutions/jinja.py
